# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for URL Shortener",
        "title": "URL Shortener",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/api/v1/analytics/{short_url}": {
            "get": {
                "description": "get analytics by provided short url",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ANALYTICS API"
                ],
                "summary": "get analytics by short url",
                "parameters": [
                    {
                        "type": "string",
                        "description": "short url",
                        "name": "short_url",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AnAgregation"
                        }
                    },
                    "400": {
                        "description": "invalid short url",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/s/{short_url}": {
            "get": {
                "description": "redirect to original url by provided short url",
                "produces": [
                    "application/json",
                    "text/plain",
                    "text/html"
                ],
                "tags": [
                    "URL API"
                ],
                "summary": "redirect by short url",
                "parameters": [
                    {
                        "type": "string",
                        "description": "short url",
                        "name": "short_url",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "302": {
                        "description": "redirect by short url",
                        "headers": {
                            "Location": {
                                "type": "string",
                                "description": "redirect URL"
                            }
                        }
                    },
                    "400": {
                        "description": "invalid short url",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "provided short url does not exist",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/shorten": {
            "post": {
                "description": "create short url with provided params",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "URL API"
                ],
                "summary": "create short url",
                "parameters": [
                    {
                        "description": "create short url request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.CreateShortURLRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.CreateShortURLResponse"
                        }
                    },
                    "400": {
                        "description": "validate error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "409": {
                        "description": "conflict error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.AnAgregation": {
            "type": "object",
            "properties": {
                "day_count": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "month_count": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "short_url": {
                    "type": "string"
                },
                "total_count": {
                    "type": "integer"
                },
                "user_agent_count": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                }
            }
        },
        "request.CreateShortURLRequest": {
            "type": "object",
            "required": [
                "url"
            ],
            "properties": {
                "short_url": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "response.CreateShortURLResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "short_url": {
                    "type": "string"
                }
            }
        }
    }
}